refactor(game): extract undo.ts, drop as-any casts (Phase 2 of #132/#96)
Move undoLastInActivePlay and canUndoActivePlay into src/Game/undo.ts as
free functions. Now that types promotes undo.frames to a typed
BackgammonGameMoving[] on BasePlay, the functions read
game.activePlay.undo.frames directly with no 'as any' casts (removes 5
casts from the undo slice of #96). The public Game.* statics remain as
thin delegating wrappers; canUndoActivePlay (deferred from guards) lands
here with its concern.

Depends on backgammon-types refactor/undo-frames-typed.
Characterization tests (10) added in the prior commit stay green.

No behavior change. Core: 458 passed / 12 skipped, tsc -b clean.

Refs #132, #96
