+++ src/Game/index.ts
@@ -44,6 +44,7 @@ import {
   canRollForStart,
 } from './guards'
 import { createBaseGameProperties, incrementStateVersion } from './shared'
+import { canUndoActivePlay, undoLastInActivePlay } from './undo'
 
 export * from '../index'
 // Import tuple aliases from types package
@@ -2418,31 +2419,16 @@ export class Game {
    * Undo the last executed move within the current activePlay using the turn-local undo stack.
    * Returns the exact pre-move moving game state.
    */
-  public static undoLastInActivePlay = function undoLastInActivePlay(
+  public static undoLastInActivePlay = function (
     game: BackgammonGame
   ): BackgammonGameMoving {
-    if (!game) throw new Error('No game state provided')
-    if (game.stateKind !== 'moving' && game.stateKind !== 'moved') {
-      throw new Error(`Cannot undo in ${game.stateKind} state. Must be in 'moving' or 'moved'`)
-    }
-    const ap: any = (game as any).activePlay
-    if (!ap) throw new Error('No active play found for undo')
-    const frames: any[] | undefined = ap.undo?.frames
-    if (!frames || frames.length === 0) throw new Error('No moves to undo for current player')
-    const previous = frames.pop()
-    if (!previous || previous.stateKind !== 'moving') throw new Error('Undo snapshot is invalid or not a moving state')
-    return previous as BackgammonGameMoving
+    return undoLastInActivePlay(game)
   }
 
   /**
    * Game-level check for whether an undo is currently possible within activePlay.
    */
-  public static canUndoActivePlay = function canUndoActivePlay(game: BackgammonGame): boolean {
-    if (!game) return false
-    if (game.stateKind !== 'moving' && game.stateKind !== 'moved') return false
-    const ap: any = (game as any).activePlay
-    if (!ap || !ap.undo) return false
-    const frames: any[] | undefined = ap.undo.frames
-    return Array.isArray(frames) && frames.length > 0
+  public static canUndoActivePlay = function (game: BackgammonGame): boolean {
+    return canUndoActivePlay(game)
   }
 }
